# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for Trackerr service.",
        "title": "Trackerr",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.banjo.dev:8080",
    "basePath": "/api/v1",
    "paths": {
        "/models": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a list of all tracker models currently supported",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Get list of tracker models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Model"
                            }
                        }
                    },
                    "400": {
                        "description": "API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add support for new model, by specifing which SMS messages should be sent when the tracker model is provisioned. The tracker model, must support GT06 or JT808, to work.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Create model",
                "parameters": [
                    {
                        "description": "Register model payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateModelReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "400": {
                        "description": "failed to parse OR API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key OR You don't have access to this feature",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "500": {
                        "description": "failed",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            }
        },
        "/models/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Get model by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Model"
                        }
                    },
                    "400": {
                        "description": "API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "404": {
                        "description": "Model was not found",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Remove support for tracker model",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "Delete model",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Model name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "400": {
                        "description": "API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key OR You don't have access to this feature",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "500": {
                        "description": "failed",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            }
        },
        "/trackers": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "If the user is a admin, it will respond with a list of all trackers in the system, and if the user is a regular user, it will return all trackers owned by the user,",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trackers"
                ],
                "summary": "Get list of trackers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.TrackerResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Register tracking by supplying all properties of a tracker",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Trackers"
                ],
                "summary": "Register tracker",
                "parameters": [
                    {
                        "description": "Register tracker payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RegisterTrackerReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "400": {
                        "description": "failed to parse OR API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "403": {
                        "description": "This action requires admin permissions",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "409": {
                        "description": "Tracker with identical id or name already exists",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "500": {
                        "description": "Failed",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            }
        },
...
        "/trackers/{id}/command": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Send upstream command to specified tracker, and get tracker response. The request will fail if the tracker is not currently connected. Additionally the request may timeout, if the tracker is connected but does not response. This can happen if the tracker has entered sleep mode without first closing the TCP connection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Commands"
                ],
                "summary": "Send command",
                "parameters": [
                    {
                        "type": "string",
                        "description": "TrackerID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Command",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CommandReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "RESPONSE",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "400": {
                        "description": "failed to parse OR API key required",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "401": {
                        "description": "Invalid API key OR not allowed to access tracker",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    },
                    "503": {
                        "description": "The tracker is not connected",
                        "schema": {
                            "$ref": "#/definitions/api.StringResultRes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.CommandReq": {
            "type": "object",
            "required": [
                "command"
            ],
            "properties": {
                "command": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
        "api.CreateModelReq": {
            "type": "object",
            "required": [
                "init_commands",
                "name",
                "success_keywords"
            ],
            "properties": {
                "init_commands": {
                    "type": "string",
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "minLength": 2
                },
                "success_keywords": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "api.StringResultRes": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        },
...
        "model.Model": {
            "type": "object",
            "properties": {
                "init_commands": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "success_keywords": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}